+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 ---
 
 - [Optimize: remove discovery module useless code](https://github.com/Tencent/spring-cloud-tencent/pull/597)
+- [Fix issue 579:Report the labels in request when report the result of invocation by RestTemplate](https://github.com/Tencent/spring-cloud-tencent/pull/600)